refactor: move extractSolutionsFromDT method
