+++ src/main/scala/satify/update/dpll/DPLL.scala
@@ -16,27 +16,24 @@ object DPLL:
 
   private val rnd = Random(42)
 
-  /**
-   * DPLL (Davis-Putnam-Logemann-Loveland) algorithm.
-   * @param dec Decision to apply
-   * @return built DecisionTree from dec
-   */
+  /** DPLL (Davis-Putnam-Logemann-Loveland) algorithm.
+    * @param dec Decision to apply
+    * @return built DecisionTree from dec
+    */
   def dpll(dec: Decision): DecisionTree =
 
-    /**
-     * Recursively call dpll applying the given constraint.
-     * @param dec the previous Decision
-     * @param constr Constraint to apply
-     * @return updated DecisionTree with the new Decision
-     */
+    /** Recursively call dpll applying the given constraint.
+      * @param dec the previous Decision
+      * @param constr Constraint to apply
+      * @return updated DecisionTree with the new Decision
+      */
     def decide(dec: Decision, constr: Constraint): DecisionTree = dec match
       case Decision(parModel, cnf) => dpll(Decision(updateParModel(parModel, constr), simplifyCnf(cnf, constr)))
 
-    /**
-     * Branch on a unit literal.
-     * @param dec the previous Decision
-     * @return an Option containing a Branch if there's a unit literal, None otherwise.
-     */
+    /** Branch on a unit literal.
+      * @param dec the previous Decision
+      * @return an Option containing a Branch if there's a unit literal, None otherwise.
+      */
     def unitPropBranch(dec: Decision): Option[Branch] = dec match
       case Decision(parModel, cnf) =>
         unitPropagation(cnf) match
@@ -50,12 +47,11 @@ object DPLL:
             )
           case None => None
 
-    /**
-     * Branch the decision tree by choosing a random variable among the available ones and by
-     * applying a random constraint.
-     * @param dec the previous Decision
-     * @return updated DecisionTree with the random Decision.
-     */
+    /** Branch the decision tree by choosing a random variable among the available ones and by
+      * applying a random constraint.
+      * @param dec the previous Decision
+      * @return updated DecisionTree with the random Decision.
+      */
     def randomBranch(dec: Decision): DecisionTree = dec match
       case Decision(parModel, cnf) =>
         val unVars = filterUnconstrVars(extractModelFromCnf(cnf))
@@ -78,23 +74,10 @@ object DPLL:
             case Some(branch) => branch
             case None => randomBranch(dec)
 
-  /** Get all SAT solutions, e.g. all Leaf nodes where the CNF has been simplified to Symbol(True).
-    * @param dt DecisionTree
-    * @return a set of PartialModel(s).
+  /** Apply unit propagation.
+    * @param cnf where to search for a unit literal
+    * @return an Option containing a Constraint if a unit literal is present, None otherwise.
     */
-  def extractSolutionsFromDT(dt: DecisionTree): Set[PartialModel] =
-    dt match
-      case Leaf(Decision(pm, cnf)) =>
-        cnf match
-          case Symbol(True) => Set(pm)
-          case _ => Set.empty
-      case Branch(_, left, right) => extractSolutionsFromDT(left) ++ extractSolutionsFromDT(right)
-
-  /**
-   * Apply unit propagation.
-   * @param cnf where to search for a unit literal
-   * @return an Option containing a Constraint if a unit literal is present, None otherwise.
-   */
   private def unitPropagation(cnf: CNF): Option[Constraint] =
 
     val f: (CNF, Option[Constraint]) => Option[Constraint] = (cnf, d) =>
