+++ src/test/scala/satify/DPLLTest/PartialModelUtilsTest.scala
@@ -31,6 +31,17 @@ class PartialModelUtilsTest extends AnyFlatSpec with Matchers:
     updateParModel(parModel, Constraint("b", false)) shouldBe Seq(varA, Variable("b", Some(false)))
   }
 
+  "All solutions" should "be extractable from a DecisionTree" in {
+    extractSolutionsFromDT(dpll(Decision(extractModelFromCnf(cnf), cnf))) shouldBe
+      Set(Seq(Variable("a", Some(true)), Variable("b", Some(true))))
+    Dpll(And(Symbol(varA), Or(Symbol(varB), Symbol(varC)))) shouldBe
+      Set(
+        Seq(Variable("a", Some(true)), Variable("b", Some(true)), Variable("c", Some(true))),
+        Seq(Variable("a", Some(true)), Variable("b", Some(true)), Variable("c", Some(false))),
+        Seq(Variable("a", Some(true)), Variable("b", Some(false)), Variable("c", Some(true)))
+      )
+  }
+
   "All solutions" should "be extractable from a PartialModel" in {
     val pm: PartialModel = Seq(Variable("a"), Variable("b"), Variable("c", Some(true)))
     explodeSolutions(pm) shouldBe
